8288537: Refactor hotspot/share/memory/iterator.hpp
