+++ src/hotspot/share/cds/metaspaceShared.hpp
@@ -34,6 +34,7 @@
 
 class FileMapInfo;
 class outputStream;
+class SerializeClosure;
 
 template<class E> class GrowableArray;
