+++ src/hotspot/share/classfile/vmSymbols.hpp
@@ -32,6 +32,8 @@
 #include "utilities/macros.hpp"
 #include "utilities/enumIterator.hpp"
 
+class SerializeClosure;
+
 // The class vmSymbols is a name space for fast lookup of
 // symbols commonly used in the VM.
 //
