+++ src/hotspot/share/oops/instanceStackChunkKlass.cpp
@@ -23,10 +23,11 @@
  */
 
 #include "precompiled.hpp"
+#include "cds/serializeClosure.hpp"
 #include "classfile/javaClasses.inline.hpp"
 #include "classfile/vmClasses.hpp"
 #include "compiler/oopMap.inline.hpp"
-#include "memory/iterator.inline.hpp"
+#include "memory/devirtualizer.inline.hpp"
 #include "memory/oopFactory.hpp"
 #include "memory/resourceArea.hpp"
 #include "oops/instanceStackChunkKlass.inline.hpp"
