+++ src/hotspot/share/oops/instanceStackChunkKlass.inline.hpp
@@ -35,7 +35,7 @@
 #include "gc/shared/collectedHeap.hpp"
 #include "gc/shared/gc_globals.hpp"
 #include "logging/log.hpp"
-#include "memory/iterator.inline.hpp"
+#include "memory/devirtualizer.inline.hpp"
 #include "oops/instanceKlass.inline.hpp"
 #include "oops/klass.hpp"
 #include "oops/oop.inline.hpp"
